Fixed map resizing when sider is collapsed. Added AutoResizeMapComponent, centralized MapContext.Provider in App component.
