+++ frontend/src/App.tsx
@@ -1,4 +1,5 @@
 import { ApiOutlined, GithubOutlined } from '@ant-design/icons';
+import { MapContext } from '@terrestris/react-geo';
 import { Layout, Space } from 'antd';
 import React, { useState } from 'react';
 import { BrowserRouter as Router, Navigate, Outlet, Route, Routes, useLocation } from 'react-router-dom';
@@ -23,6 +24,7 @@ import { useAuth } from './Hooks/useAuth';
 import CatalogueServiceRepo from './Repos/CswRepo';
 import WebFeatureServiceRepo from './Repos/WfsRepo';
 import WebMapServiceRepo from './Repos/WmsRepo';
+import { olMap } from './Utils/MapUtils';
 
 
 const { Content, Footer, Sider } = Layout;
@@ -61,37 +63,39 @@ export default function App (): JSX.Element {
           path='/'
           element={
             <RequireAuth>
-              <Layout style={{ minHeight: '100vh' }}>
-                <Sider
-                  collapsible
-                  collapsed={collapsed}
-                  onCollapse={onCollapse}>
-                  <div className='logo'>
-                    <img
-                      src={process.env.PUBLIC_URL + '/logo.png'}
-                      alt='Mr. Map Logo'
-                    >
-                    </img>
-                  </div>
-                  <NavMenu />
-                </Sider>
-                <Layout className='site-layout'>
-                  <Content style={{ margin: '0 16px' }}>
-                    <div
-                      className='site-layout-background'
-                      style={{ padding: 24, minHeight: 360 }}
-                    >
-                      <Outlet />
+              <MapContext.Provider value={olMap}>
+                <Layout style={{ minHeight: '100vh' }}>
+                  <Sider
+                    collapsible
+                    collapsed={collapsed}
+                    onCollapse={onCollapse}>
+                    <div className='logo'>
+                      <img
+                        src={process.env.PUBLIC_URL + '/logo.png'}
+                        alt='Mr. Map Logo'
+                      >
+                      </img>
                     </div>
-                  </Content>
-                  <Footer style={{ textAlign: 'center' }}>
-                    <Space>
-                      <a href={swaggerUiUrl}><ApiOutlined /> OpenAPI</a>
-                      <a href='https://github.com/mrmap-community/mrmap'><GithubOutlined /> GitHub</a>
-                    </Space>
-                  </Footer>
+                    <NavMenu />
+                  </Sider>
+                  <Layout className='site-layout'>
+                    <Content style={{ margin: '0 16px' }}>
+                      <div
+                        className='site-layout-background'
+                        style={{ padding: 24, minHeight: 360 }}
+                      >
+                        <Outlet />
+                      </div>
+                    </Content>
+                    <Footer style={{ textAlign: 'center' }}>
+                      <Space>
+                        <a href={swaggerUiUrl}><ApiOutlined /> OpenAPI</a>
+                        <a href='https://github.com/mrmap-community/mrmap'><GithubOutlined /> GitHub</a>
+                      </Space>
+                    </Footer>
+                  </Layout>
                 </Layout>
-              </Layout>
+              </MapContext.Provider>
             </RequireAuth>
           }
         >
